refactor: Updating upgrade flows to start using upgrade hook for better code reusability (#18735)
* updated upgrade flow for authentication page to start using upgrade hook

* using the upgrade hook on upgrade pages as well

* adding RBAC feature flag
