+++ app/server/appsmith-server/src/main/resources/features/init-flags.yml
@@ -70,3 +70,13 @@ ff4j:
         param:
           - name: emailDomains
             value: appsmith.com,moolya.com
+
+    # Put EE flags below this line, to avoid conflicts.
+    - uid: RBAC
+      enable: true
+      description: Enable access control settings
+      flipstrategy:
+        class: com.appsmith.server.featureflags.strategies.EmailBasedRolloutStrategy
+        param:
+          - name: emailDomains
+            value: appsmith.com
